Added tuple header just in case that's the problem

+++ src/InputHandling/Data.h
@@ -1,12 +1,12 @@
 #pragma once
-#include <vector>
+#include <tuple>
 #include <vector>
 
 #include "../settings.h"
 #include "../Utility/Log.h"
 
-typedef unsigned int lint;
-typedef unsigned int dnaindex; 
+typedef unsigned long long int lint;
+typedef unsigned long long int dnaindex; 
 
 class Datum
 {
